Refactored Parquet test setup
